+++ src/main/java/ru/msu/cmc/webprac/dao/RentDao.java
@@ -5,7 +5,7 @@ import ru.msu.cmc.webprac.models.Rent;
 import java.sql.Timestamp;
 import java.util.List;
 
-public interface RentDao extends CommonDao<Rent, Long> {
+public interface RentDao extends CommonDao<Rent, Integer> {
 
     List<Rent> getAllRentByPeriod(Timestamp from, Timestamp to);
 }
